Increase idle DB connection timeout to 2 hours (#158)
I discovered that the connections used to acquire our database locks
were being timed out after 60 seconds and closed, because they were idle
- no queries are run on them while the lock is held.

When the connection is closed the locks are released and another task
can run concurrently :(

This is a mitigation rather than a fix. In hindsight I was wrong to
implement a custom locking solution and should have used an off the
shelf job framework, which should be the long term plan.